Cardano.Binary.Annotated

Documentation

data Annotated b a

Constructors

Annotated

Fields

unAnnotated ∷ !b

annotation ∷ !a

data ByteSpan

A pair of offsets delimiting the beginning and end of a substring of a ByteString

Constructors

ByteSpan !ByteOffset !ByteOffset

class Decoded t where

Associated Types

type BaseType t ∷ Type

Methods

recoverBytes ∷ t → ByteString

annotationBytes ∷ Functor f ⇒ LByteString → f ByteSpan → f ByteString

annotatedDecoder ∷ Decoder s a → Decoder s (Annotated a ByteSpan)

A decoder for a value paired with an annotation specifying the start and end of the consumed bytes.

slice ∷ ByteString → ByteSpan → LByteString

Extract a substring of a given ByteString corresponding to the offsets.

fromCBORAnnotated ∷ FromCBOR a ⇒ Decoder s (Annotated a ByteSpan)

A decoder for a value paired with an annotation specifying the start and end of the consumed bytes.

decodeFullAnnotatedBytes ∷ Functor f ⇒ Text → (∀ s. Decoder s (f ByteSpan)) → LByteString → Either DecoderError (f ByteString)

Decodes a value from a ByteString, requiring that the full ByteString is consumed, and replaces ByteSpan annotations with the corresponding substrings of the input string.

reAnnotate ∷ ToCBOR a ⇒ Annotated a b → Annotated a ByteString

Reconstruct an annotation by re-serialising the payload to a ByteString.

type AnnotatedDecoder s a = ReaderT LByteString (Decoder s) a

An AnnotatedDecoder produces a value which needs a reference to the original ByteString to be constructed. For example, consider

`data Foo = Foo Int ByteString`

where the ByteString is expected to be the serialized form of Foo. A Decoder for Foo would need to duplicate the bytes as it read them in order to populate the ByteString. With an AnnotatedDecoder we instead fill the ByteString using the reader.

decodeFooAnnotated = withSlice $ do int <- lift decodeInt bytes <- get pure (Foo int bytes)

This assumes that the ByteString passed into the reader will be the same as the ByteString passed to the decoder.

class FromCBORAnnotated a where

Methods

fromCBORAnnotated' ∷ AnnotatedDecoder s a

decodeAnnotated ∷ ∀ a. (Typeable a, FromCBORAnnotated a) ⇒ LByteString → Either DecoderError a

withSlice ∷ AnnotatedDecoder s (LByteString → a) → AnnotatedDecoder s a

Inserts the decoded segment

withSlice' ∷ AnnotatedDecoder s (ByteString → a) → AnnotatedDecoder s a

Equivalent to withSlice for strict ByteStrings

liftByteSpanDecoder ∷ Functor f ⇒ Decoder s (f ByteSpan) → AnnotatedDecoder s (f ByteString)

decodeAnnotatedDecoder ∷ Text → (∀ s. AnnotatedDecoder s a) → LByteString → Either DecoderError a

fromCBOREmptyAnnotation ∷ FromCBORAnnotated a ⇒ Decoder s a