Cardano.Binary.Deserialize

Description

Deserialization primitives built on top of the FromCBOR typeclass

Unsafe deserialization

unsafeDeserialize ∷ FromCBOR a ⇒ ByteString → a

Deserialize a Haskell value from the external binary representation (which must have been made using serialize or related function).

Throws: DeserialiseFailure if the given external representation is invalid or does not correspond to a value of the expected type.

unsafeDeserialize' ∷ FromCBOR a ⇒ ByteString → a

Strict variant of deserialize.

toStrictByteString

Arguments

∷ Encoding
The Encoding of a CBOR value.

→ ByteString
The encoded value.

Turn an Encoding into a strict ByteString in CBOR binary format.

Since: cborg-0.2.0.0

Decoding

decodeFull ∷ ∀ a. FromCBOR a ⇒ ByteString → Either DecoderError a

Deserialize a Haskell value from the external binary representation, failing if there are leftovers. In a nutshell, the full here implies the contract of this function is that what you feed as input needs to be consumed entirely.

decodeFull' ∷ ∀ a. FromCBOR a ⇒ ByteString → Either DecoderError a

decodeFullDecoder

Arguments

∷ Text
Label for error reporting

→ (∀ s. Decoder s a)
The parser for the ByteString to decode. It should decode the given ByteString into a value of type a

→ ByteString
The ByteString to decode

→ Either DecoderError a

decodeFullDecoder'

Arguments

∷ Text
Label for error reporting

→ (∀ s. Decoder s a)
The parser for the ByteString to decode. It should decode the given ByteString into a value of type a

→ ByteString
The ByteString to decode

→ Either DecoderError a

CBOR in CBOR

decodeNestedCbor ∷ FromCBOR a ⇒ Decoder s a

Remove the the semantic tag 24 from the enclosed CBOR data item, decoding back the inner ByteString as a proper Haskell type. Consume its input in full.

decodeNestedCborBytes ∷ Decoder s ByteString

Like decodeKnownCborDataItem, but assumes nothing about the Haskell type we want to deserialise back, therefore it yields the ByteString Tag 24 surrounded (stripping such tag away).

In CBOR notation, if the data was serialised as:

>>> 24(h'DEADBEEF')

then decodeNestedCborBytes yields the inner DEADBEEF, unchanged.