Cardano.Binary.FromCBOR

Documentation

class Typeable a ⇒ FromCBOR a where

Minimal complete definition

fromCBOR

Methods

fromCBOR ∷ Decoder s a

label ∷ Proxy a → Text

data DecoderError

Constructors

DecoderErrorCanonicityViolation Text

DecoderErrorCustom Text Text
Custom decoding error, usually due to some validation failure

DecoderErrorDeserialiseFailure Text DeserialiseFailure

DecoderErrorEmptyList Text

DecoderErrorLeftover Text ByteString

DecoderErrorSizeMismatch Text Int Int
A size mismatch DecoderErrorSizeMismatch label expectedSize actualSize

DecoderErrorUnknownTag Text Word8

DecoderErrorVoid

enforceSize ∷ Text → Int → Decoder s ()

Enforces that the input size is the same as the decoded one, failing in case it's not

matchSize ∷ Text → Int → Int → Decoder s ()

Compare two sizes, failing if they are not equal

fromCBORMaybe ∷ Decoder s a → Decoder s (Maybe a)

decodeListWith ∷ Decoder s a → Decoder s [a]

Decoder for list.