Cardano.Binary.Serialize

Description

Serialization primitives built on top of the ToCBOR typeclass

Documentation

serialize ∷ ToCBOR a ⇒ a → ByteString

Serialize a Haskell value with a ToCBOR instance to an external binary representation.

The output is represented as a lazy LByteString and is constructed incrementally.

serialize' ∷ ToCBOR a ⇒ a → ByteString

Serialize a Haskell value to an external binary representation.

The output is represented as a strict ByteString.

serializeBuilder ∷ ToCBOR a ⇒ a → Builder

Serialize into a Builder. Useful if you want to throw other ByteStrings around it.

serializeEncoding ∷ Encoding → ByteString

Deprecated: Use serialize instead, since Encoding has ToCBOR instance

Serialize a Haskell value to an external binary representation using the provided CBOR Encoding

The output is represented as an LByteString and is constructed incrementally.

serializeEncoding' ∷ Encoding → ByteString

Deprecated: Use serialize` instead, since Encoding has ToCBOR instance

A strict version of serializeEncoding

CBOR in CBOR

encodeNestedCbor ∷ ToCBOR a ⇒ a → Encoding

Encode and serialise the given a and sorround it with the semantic tag 24 In CBOR diagnostic notation: >>> 24(hDEADBEEF)

encodeNestedCborBytes ∷ ByteString → Encoding

Like encodeNestedCbor, but assumes nothing about the shape of input object, so that it must be passed as a binary ByteString blob. It's the caller responsibility to ensure the input ByteString correspond indeed to valid, previously-serialised CBOR data.

nestedCborSizeExpr ∷ Size → Size

nestedCborBytesSizeExpr ∷ Size → Size