Cardano.Binary

Documentation

class Typeable a ⇒ ToCBOR a where

Minimal complete definition

toCBOR

Methods

toCBOR ∷ a → Encoding

encodedSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy a → Size

encodedListSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy [a] → Size

withWordSize ∷ (Integral s, Integral a) ⇒ s → a

Compute encoded size of an integer

module Codec.CBOR.Encoding

encodeMaybe ∷ (a → Encoding) → Maybe a → Encoding

toCBORMaybe ∷ (a → Encoding) → Maybe a → Encoding

Deprecated: In favor of encodeMaybe

encodeNullMaybe ∷ (a → Encoding) → Maybe a → Encoding

Alternative way to encode a Maybe type.

Note - this is not the default method for encoding Maybe, use encodeMaybe instead

encodeSeq ∷ (a → Encoding) → Seq a → Encoding

encodeNominalDiffTime ∷ NominalDiffTime → Encoding

encodeNominalDiffTimeMicro ∷ NominalDiffTime → Encoding

Same as encodeNominalDiffTime, except with loss of precision, because it encoded as Micro

Size of expressions

data Range b

A range of values. Should satisfy the invariant forall x. lo x <= hi x.

Constructors

Range

Fields

lo ∷ b

hi ∷ b

szEval ∷ (∀ t. ToCBOR t ⇒ (Proxy t → Size) → Proxy t → Range Natural) → Size → Range Natural

Fully evaluate a size expression by applying the given function to any suspended computations. szEval g effectively turns each "thunk" of the form TodoF f x into g x, then evaluates the result.

type Size = Fix SizeF

Expressions describing the statically-computed size bounds on a type's possible values.

data Case t

An individual labeled case.

Constructors

Case Text t

caseValue ∷ Case t → t

Discard the label on a case.

newtype LengthOf xs

A type used to represent the length of a value in Size computations.

Constructors

LengthOf xs

data SizeOverride

Override mechanisms to be used with szWithCtx.

Constructors

SizeConstant Size
Replace with a fixed Size.

SizeExpression ((∀ a. ToCBOR a ⇒ Proxy a → Size) → Size)
Recursively compute the size.

SelectCases [Text]
Select only a specific case from a CasesF.

isTodo ∷ Size → Bool

Is this expression a thunk?

szCases ∷ [Case Size] → Size

Create a case expression from individual cases.

szLazy ∷ ToCBOR a ⇒ Proxy a → Size

Evaluate the expression lazily, by immediately creating a thunk that will evaluate its contents lazily.

ghci> putStrLn $ pretty $ szLazy (Proxy @TxAux)
(_ :: TxAux)

szGreedy ∷ ToCBOR a ⇒ Proxy a → Size

Evaluate an expression greedily. There may still be thunks in the result, for types that did not provide a custom encodedSizeExpr method in their ToCBOR instance.

ghci> putStrLn $ pretty $ szGreedy (Proxy @TxAux)
(0 + { TxAux=(2 + ((0 + (((1 + (2 + ((_ :: LengthOf [TxIn]) * (2 + { TxInUtxo=(2 + ((1 + 34) + { minBound=1 maxBound=5 })) })))) + (2 + ((_ :: LengthOf [TxOut]) * (0 + { TxOut=(2 + ((0 + ((2 + ((2 + withWordSize((((1 + 30) + (_ :: Attributes AddrAttributes)) + 1))) + (((1 + 30) + (_ :: Attributes AddrAttributes)) + 1))) + { minBound=1 maxBound=5 })) + { minBound=1 maxBound=9 })) })))) + (_ :: Attributes ()))) + (_ :: Vector TxInWitness))) })

szForce ∷ Size → Size

Force any thunks in the given Size expression.

ghci> putStrLn $ pretty $ szForce $ szLazy (Proxy @TxAux)
(0 + { TxAux=(2 + ((0 + (_ :: Tx)) + (_ :: Vector TxInWitness))) })

szWithCtx ∷ ToCBOR a ⇒ Map TypeRep SizeOverride → Proxy a → Size

Greedily compute the size bounds for a type, using the given context to override sizes for specific types.

szSimplify ∷ Size → Either Size (Range Natural)

Simplify the given Size, resulting in either the simplified Size or, if it was fully simplified, an explicit upper and lower bound.

apMono ∷ Text → (Natural → Natural) → Size → Size

Apply a monotonically increasing function to the expression. There are three cases when applying f to a Size expression: * When applied to a value x, compute f x. * When applied to cases, apply to each case individually. * In all other cases, create a deferred application of f.

szBounds ∷ ToCBOR a ⇒ a → Either Size (Range Natural)

serialize ∷ ToCBOR a ⇒ a → ByteString

Serialize a Haskell value with a ToCBOR instance to an external binary representation.

The output is represented as a lazy LByteString and is constructed incrementally.

serialize' ∷ ToCBOR a ⇒ a → ByteString

Serialize a Haskell value to an external binary representation.

The output is represented as a strict ByteString.

serializeBuilder ∷ ToCBOR a ⇒ a → Builder

Serialize into a Builder. Useful if you want to throw other ByteStrings around it.

serializeEncoding ∷ Encoding → ByteString

Deprecated: Use serialize instead, since Encoding has ToCBOR instance

Serialize a Haskell value to an external binary representation using the provided CBOR Encoding

The output is represented as an LByteString and is constructed incrementally.

serializeEncoding' ∷ Encoding → ByteString

Deprecated: Use serialize` instead, since Encoding has ToCBOR instance

A strict version of serializeEncoding

CBOR in CBOR

encodeNestedCbor ∷ ToCBOR a ⇒ a → Encoding

Encode and serialise the given a and sorround it with the semantic tag 24 In CBOR diagnostic notation: >>> 24(hDEADBEEF)

encodeNestedCborBytes ∷ ByteString → Encoding

Like encodeNestedCbor, but assumes nothing about the shape of input object, so that it must be passed as a binary ByteString blob. It's the caller responsibility to ensure the input ByteString correspond indeed to valid, previously-serialised CBOR data.

nestedCborSizeExpr ∷ Size → Size

nestedCborBytesSizeExpr ∷ Size → Size

class Typeable a ⇒ FromCBOR a where

Minimal complete definition

fromCBOR

Methods

fromCBOR ∷ Decoder s a

label ∷ Proxy a → Text

data DecoderError

Constructors

DecoderErrorCanonicityViolation Text

DecoderErrorCustom Text Text
Custom decoding error, usually due to some validation failure

DecoderErrorDeserialiseFailure Text DeserialiseFailure

DecoderErrorEmptyList Text

DecoderErrorLeftover Text ByteString

DecoderErrorSizeMismatch Text Int Int
A size mismatch DecoderErrorSizeMismatch label expectedSize actualSize

DecoderErrorUnknownTag Text Word8

DecoderErrorVoid

enforceSize ∷ Text → Int → Decoder s ()

Enforces that the input size is the same as the decoded one, failing in case it's not

matchSize ∷ Text → Int → Int → Decoder s ()

Compare two sizes, failing if they are not equal

module Codec.CBOR.Decoding

decodeMaybe ∷ Decoder s a → Decoder s (Maybe a)

fromCBORMaybe ∷ Decoder s a → Decoder s (Maybe a)

Deprecated: In favor of decodeMaybe

decodeNullMaybe ∷ Decoder s a → Decoder s (Maybe a)

decodeSeq ∷ Decoder s a → Decoder s (Seq a)

decodeListWith ∷ Decoder s a → Decoder s [a]

Decoder for list.

decodeNominalDiffTime ∷ Decoder s NominalDiffTime

decodeNominalDiffTimeMicro ∷ Decoder s NominalDiffTime

For backwards compatibility we round pico precision to micro

Helper tools to build instances

decodeMapSkel ∷ (Ord k, FromCBOR k, FromCBOR v) ⇒ ([(k, v)] → m) → Decoder s m

Checks canonicity by comparing the new key being decoded with the previous one, to enfore these are sorted the correct way. See: https://tools.ietf.org/html/rfc7049#section-3.9 "[..]The keys in every map must be sorted lowest value to highest.[...]"

decodeCollection ∷ Decoder s (Maybe Int) → Decoder s a → Decoder s [a]

decodeCollectionWithLen ∷ Decoder s (Maybe Int) → Decoder s v → Decoder s (Int, [v])

cborError ∷ (MonadFail m, Buildable e) ⇒ e → m a

Convert a Buildable error message into a MonadFail failure.

toCborError ∷ (MonadFail m, Buildable e) ⇒ Either e a → m a

Convert an Either-encoded failure to a MonadFail failure using the Buildable insatance

Unsafe deserialization

unsafeDeserialize ∷ FromCBOR a ⇒ ByteString → a

Deserialize a Haskell value from the external binary representation (which must have been made using serialize or related function).

Throws: DeserialiseFailure if the given external representation is invalid or does not correspond to a value of the expected type.

unsafeDeserialize' ∷ FromCBOR a ⇒ ByteString → a

Strict variant of deserialize.

toStrictByteString

Arguments

∷ Encoding
The Encoding of a CBOR value.

→ ByteString
The encoded value.

Turn an Encoding into a strict ByteString in CBOR binary format.

Since: cborg-0.2.0.0

Decoding

decodeFull ∷ ∀ a. FromCBOR a ⇒ ByteString → Either DecoderError a

Deserialize a Haskell value from the external binary representation, failing if there are leftovers. In a nutshell, the full here implies the contract of this function is that what you feed as input needs to be consumed entirely.

decodeFull' ∷ ∀ a. FromCBOR a ⇒ ByteString → Either DecoderError a

decodeFullDecoder

Arguments

∷ Text
Label for error reporting

→ (∀ s. Decoder s a)
The parser for the ByteString to decode. It should decode the given ByteString into a value of type a

→ ByteString
The ByteString to decode

→ Either DecoderError a

decodeFullDecoder'

Arguments

∷ Text
Label for error reporting

→ (∀ s. Decoder s a)
The parser for the ByteString to decode. It should decode the given ByteString into a value of type a

→ ByteString
The ByteString to decode

→ Either DecoderError a

CBOR in CBOR

decodeNestedCbor ∷ FromCBOR a ⇒ Decoder s a

Remove the the semantic tag 24 from the enclosed CBOR data item, decoding back the inner ByteString as a proper Haskell type. Consume its input in full.

decodeNestedCborBytes ∷ Decoder s ByteString

Like decodeKnownCborDataItem, but assumes nothing about the Haskell type we want to deserialise back, therefore it yields the ByteString Tag 24 surrounded (stripping such tag away).

In CBOR notation, if the data was serialised as:

>>> 24(h'DEADBEEF')

then decodeNestedCborBytes yields the inner DEADBEEF, unchanged.