Cardano.Binary.FromCBOR

Documentation

class Typeable a ⇒ FromCBOR a where

Minimal complete definition

fromCBOR

Methods

fromCBOR ∷ Decoder s a

label ∷ Proxy a → Text

data DecoderError

Constructors

DecoderErrorCanonicityViolation Text

DecoderErrorCustom Text Text
Custom decoding error, usually due to some validation failure

DecoderErrorDeserialiseFailure Text DeserialiseFailure

DecoderErrorEmptyList Text

DecoderErrorLeftover Text ByteString

DecoderErrorSizeMismatch Text Int Int
A size mismatch DecoderErrorSizeMismatch label expectedSize actualSize

DecoderErrorUnknownTag Text Word8

DecoderErrorVoid

enforceSize ∷ Text → Int → Decoder s ()

Enforces that the input size is the same as the decoded one, failing in case it's not

matchSize ∷ Text → Int → Int → Decoder s ()

Compare two sizes, failing if they are not equal

module Codec.CBOR.Decoding

decodeMaybe ∷ Decoder s a → Decoder s (Maybe a)

fromCBORMaybe ∷ Decoder s a → Decoder s (Maybe a)

Deprecated: In favor of decodeMaybe

decodeNullMaybe ∷ Decoder s a → Decoder s (Maybe a)

decodeSeq ∷ Decoder s a → Decoder s (Seq a)

decodeListWith ∷ Decoder s a → Decoder s [a]

Decoder for list.

decodeNominalDiffTime ∷ Decoder s NominalDiffTime

decodeNominalDiffTimeMicro ∷ Decoder s NominalDiffTime

For backwards compatibility we round pico precision to micro

Helper tools to build instances

decodeMapSkel ∷ (Ord k, FromCBOR k, FromCBOR v) ⇒ ([(k, v)] → m) → Decoder s m

Checks canonicity by comparing the new key being decoded with the previous one, to enfore these are sorted the correct way. See: https://tools.ietf.org/html/rfc7049#section-3.9 "[..]The keys in every map must be sorted lowest value to highest.[...]"

decodeCollection ∷ Decoder s (Maybe Int) → Decoder s a → Decoder s [a]

decodeCollectionWithLen ∷ Decoder s (Maybe Int) → Decoder s v → Decoder s (Int, [v])

cborError ∷ (MonadFail m, Buildable e) ⇒ e → m a

Convert a Buildable error message into a MonadFail failure.

toCborError ∷ (MonadFail m, Buildable e) ⇒ Either e a → m a

Convert an Either-encoded failure to a MonadFail failure using the Buildable insatance