Cardano.Binary.ToCBOR

Documentation

class Typeable a ⇒ ToCBOR a where

Minimal complete definition

toCBOR

Methods

toCBOR ∷ a → Encoding

encodedSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy a → Size

encodedListSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy [a] → Size

withWordSize ∷ (Integral s, Integral a) ⇒ s → a

Compute encoded size of an integer

toCBORMaybe ∷ (a → Encoding) → Maybe a → Encoding

Size of expressions

data Range b

A range of values. Should satisfy the invariant forall x. lo x <= hi x.

Constructors

Range

Fields

lo ∷ b

hi ∷ b

szEval ∷ (∀ t. ToCBOR t ⇒ (Proxy t → Size) → Proxy t → Range Natural) → Size → Range Natural

Fully evaluate a size expression by applying the given function to any suspended computations. szEval g effectively turns each "thunk" of the form TodoF f x into g x, then evaluates the result.

type Size = Fix SizeF

Expressions describing the statically-computed size bounds on a type's possible values.

data Case t

An individual labeled case.

Constructors

Case Text t

caseValue ∷ Case t → t

Discard the label on a case.

newtype LengthOf xs

A type used to represent the length of a value in Size computations.

Constructors

LengthOf xs

data SizeOverride

Override mechanisms to be used with szWithCtx.

Constructors

SizeConstant Size
Replace with a fixed Size.

SizeExpression ((∀ a. ToCBOR a ⇒ Proxy a → Size) → Size)
Recursively compute the size.

SelectCases [Text]
Select only a specific case from a CasesF.

isTodo ∷ Size → Bool

Is this expression a thunk?

szCases ∷ [Case Size] → Size

Create a case expression from individual cases.

szLazy ∷ ToCBOR a ⇒ Proxy a → Size

Evaluate the expression lazily, by immediately creating a thunk that will evaluate its contents lazily.

ghci> putStrLn $ pretty $ szLazy (Proxy @TxAux)
(_ :: TxAux)

szGreedy ∷ ToCBOR a ⇒ Proxy a → Size

Evaluate an expression greedily. There may still be thunks in the result, for types that did not provide a custom encodedSizeExpr method in their ToCBOR instance.

ghci> putStrLn $ pretty $ szGreedy (Proxy @TxAux)
(0 + { TxAux=(2 + ((0 + (((1 + (2 + ((_ :: LengthOf [TxIn]) * (2 + { TxInUtxo=(2 + ((1 + 34) + { minBound=1 maxBound=5 })) })))) + (2 + ((_ :: LengthOf [TxOut]) * (0 + { TxOut=(2 + ((0 + ((2 + ((2 + withWordSize((((1 + 30) + (_ :: Attributes AddrAttributes)) + 1))) + (((1 + 30) + (_ :: Attributes AddrAttributes)) + 1))) + { minBound=1 maxBound=5 })) + { minBound=1 maxBound=9 })) })))) + (_ :: Attributes ()))) + (_ :: Vector TxInWitness))) })

szForce ∷ Size → Size

Force any thunks in the given Size expression.

ghci> putStrLn $ pretty $ szForce $ szLazy (Proxy @TxAux)
(0 + { TxAux=(2 + ((0 + (_ :: Tx)) + (_ :: Vector TxInWitness))) })

szWithCtx ∷ ToCBOR a ⇒ Map TypeRep SizeOverride → Proxy a → Size

Greedily compute the size bounds for a type, using the given context to override sizes for specific types.

szSimplify ∷ Size → Either Size (Range Natural)

Simplify the given Size, resulting in either the simplified Size or, if it was fully simplified, an explicit upper and lower bound.

apMono ∷ Text → (Natural → Natural) → Size → Size

Apply a monotonically increasing function to the expression. There are three cases when applying f to a Size expression: * When applied to a value x, compute f x. * When applied to cases, apply to each case individually. * In all other cases, create a deferred application of f.

szBounds ∷ ToCBOR a ⇒ a → Either Size (Range Natural)